[google_maps] Split integration_test in 3 smaller files.
